requirements.txt
# --- Core scientific / plotting ---
matplotlib>=3.7
numpy>=1.25
pandas

# --- Image handling (für Logos/Icons) ---
pillow

# --- Bluetooth / Vivosun support ---
bleak
vivosun-thermo

# --- macOS integration (Dock-Icon etc.) ---
pyobjc-framework-Cocoa

# --- Build tool (optional, nur wenn selbst gebaut wird) ---
pyinstaller